extract ToDoPage
